--- src/flowcean/polars/environments/parquet.py
@@ -1,18 +1,0 @@
-from pathlib import Path
-
-import polars as pl
-
-from .dataset import Dataset
-
-
-class ParquetDataLoader(Dataset):
-    """DataLoader for Parquet files."""
-
-    def __init__(self, path: str | Path) -> None:
-        """Initialize the ParquetDataLoader.
-
-        Args:
-            path: Path to the Parquet file.
-        """
-        data = pl.scan_parquet(path)
-        super().__init__(data)
