+++ src/flowcean/ros/rosbag.py
@@ -11,7 +11,7 @@ from rosbags.interfaces import Msgdef, Nodetype
 from rosbags.typesys import Stores, get_types_from_msg, get_typestore
 from tqdm import tqdm
 
-from flowcean.polars import Dataset
+from flowcean.polars import DataFrame
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Sequence
@@ -28,7 +28,7 @@ class RosbagError(Exception):
 logger = logging.getLogger(__name__)
 
 
-class RosbagLoader(Dataset):
+class RosbagLoader(DataFrame):
     """Environment to load data from a rosbag file.
 
     The RosbagEnvironment is used to load data from a rosbag file. The
