+++ tests/polars/test_csv.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import polars as pl
 from polars.testing import assert_frame_equal
 
-from flowcean.polars import CsvDataLoader
+from flowcean.polars.environments.dataframe import DataFrame
 
 
 class TestCsvDataLoader(unittest.TestCase):
@@ -20,7 +20,7 @@ class TestCsvDataLoader(unittest.TestCase):
         with tempfile.NamedTemporaryFile() as f:
             data.write_csv(f.name)
             datapath = Path(f.name)
-            dataloader = CsvDataLoader(path=datapath)
+            dataloader = DataFrame.from_csv(path=datapath)
             loaded_data = dataloader.observe()
             assert_frame_equal(loaded_data.collect(), data)
