+++ tests/polars/test_parquet.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import polars as pl
 from polars.testing import assert_frame_equal
 
-from flowcean.polars import ParquetDataLoader
+from flowcean.polars.environments.dataframe import DataFrame
 
 
 class TestParquetDataLoader(unittest.TestCase):
@@ -20,7 +20,7 @@ class TestParquetDataLoader(unittest.TestCase):
         with tempfile.NamedTemporaryFile() as f:
             data.write_parquet(f.name)
             datapath = Path(f.name)
-            dataloader = ParquetDataLoader(path=datapath)
+            dataloader = DataFrame.from_parquet(path=datapath)
             loaded_data = dataloader.observe()
             assert_frame_equal(loaded_data.collect(), data)
